updating query store to use base class
