+++ src/Microsoft.Health.Dicom.SqlServer/Features/Query/ISqlQueryStore.cs
@@ -3,21 +3,15 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using Microsoft.Extensions.Logging;
+using Microsoft.Health.Dicom.Core.Features.Query;
 using Microsoft.Health.Dicom.SqlServer.Features.Schema;
-using Microsoft.Health.SqlServer.Features.Client;
 
 namespace Microsoft.Health.Dicom.SqlServer.Features.Query
 {
-    internal class SqlQueryStoreV2 : SqlQueryStoreV1
+    /// <summary>
+    ///  Sql version of IQueryStore.
+    /// </summary>
+    internal interface ISqlQueryStore : IQueryStore, IVersioned
     {
-        public override SchemaVersion Version => SchemaVersion.V2;
-
-        public SqlQueryStoreV2(
-            SqlConnectionWrapperFactory sqlConnectionWrapperFactory,
-            ILogger<ISqlQueryStore> logger)
-            : base(sqlConnectionWrapperFactory, logger)
-        {
-        }
     }
 }
